Upgrade from @simplewebauthn/server from 11.0.0 to 13.1.1

+++ reset-db.sh
@@ -2,4 +2,9 @@
 
 cp -n .default.env .env
 
-docker compose down -v
+if command -v docker &>/dev/null && (curl -s --unix-socket /var/run/docker.sock http/_ping 2>&1 >/dev/null || curl -s --unix-socket /run/user/$(id -u)/docker.sock http/_ping 2>&1 >/dev/null || curl -s --unix-socket ~/.colima/docker.sock http/_ping 2>&1 >/dev/null); then
+  docker compose down -v
+fi
+if command -v podman &>/dev/null; then
+  podman compose down -v
+fi
